rt-kernel: Move MIB files to new directory
This is similar to how it is done for Linux.
