+++ cmake/rt-kernel.cmake
@@ -24,11 +24,11 @@ target_sources(profinet
   src/ports/rt-kernel/pnal_eth.c
   src/ports/rt-kernel/pnal_udp.c
   src/ports/rt-kernel/pnal_snmp.c
-  src/ports/rt-kernel/mib2_system.c
-  src/ports/rt-kernel/lldp-mib.c
-  src/ports/rt-kernel/lldp-ext-pno-mib.c
-  src/ports/rt-kernel/lldp-ext-dot3-mib.c
-  src/ports/rt-kernel/rowindex.c
+  src/ports/rt-kernel/mib/mib2_system.c
+  src/ports/rt-kernel/mib/lldp-mib.c
+  src/ports/rt-kernel/mib/lldp-ext-pno-mib.c
+  src/ports/rt-kernel/mib/lldp-ext-dot3-mib.c
+  src/ports/rt-kernel/mib/rowindex.c
   src/ports/rt-kernel/dwmac1000.c
   )
